Relocate all validator file into one file
