+++ app/src/main/java/com/example/cactusnotes/validators/EmailValidator.kt
@@ -1,4 +1,6 @@
-package com.example.cactusnotes
+package com.example.cactusnotes.validators
+
+import com.example.cactusnotes.ValidationResult
 
 class EmailValidator(private val email: String) {
